Correccion Actividad con Medalla

+++ app/Http/Controllers/MedallaController.php
@@ -13,6 +13,11 @@ use App\Materia;
 use App\Ciclo;
 use App\MateriaImpartida;
 use App\MedallaGanada;
+use App\Actividad;
+use App\DetallePunto;
+use App\Grupo;
+use App\TipoActividad;
+use App\DetallePuntoActividad;
 use Laracasts\Flash\Flash;
 
 class MedallaController extends Controller
@@ -108,6 +113,7 @@ class MedallaController extends Controller
     public function store(Request $request)
     {
         //dd($request->all());
+        $grupo = Grupo::where('IDMATERIAIMPARTIDA',$request->IDMATERIAIMPARTIDA)->get();
         //Imagen
         if($request->file('imgMedalla'))
         {
@@ -117,7 +123,7 @@ class MedallaController extends Controller
           $Foto->move($path, $nombreFoto);
         }else{
           //Foto por Default
-          $nombreFoto = $MedallaFotoDefault;
+          $nombreFoto = '_MedallaDefault.png';
         }
 
         //Creacion de Medallas
@@ -141,6 +147,30 @@ class MedallaController extends Controller
         $detalleMedalla->IDMEDALLA = $medalla->id;
         //Guardar Detalle Medalla
         $detalleMedalla->save();
+
+        foreach($grupo as $g){
+            //DETALLE ACTIVIDAD NUEVA
+            $detallePunto = new DetallePunto();
+            $detallePunto->IDTIPOPUNTO = 2;
+            $detallePunto->IDMATERIAIMPARTIDA = $request->IDMATERIAIMPARTIDA;
+            $detallePunto->ESTAACTIVOPUNTO = 1;
+            $detallePunto->PUNTOSACTIVIDAD = $request->PUNTOSACTIVIDAD;
+            $detallePunto->save();
+            
+            //CREANDO TIPOACTIVIDAD
+            $tipoActividad = new TipoActividad();
+            $tipoActividad->NOMBRETIPOACTIVIDAD = $request->NOMBREMEDALLA;
+            $tipoActividad->save();
+
+            //ACTIVIDAD NUEVA
+            $actividad = new Actividad();
+            $actividad->IDTIPOACTIVIDAD = $tipoActividad->id;
+            $actividad->IDDETALLEPUNTO = $detallePunto->id;
+            $actividad->IDGRUPO = $g->id;
+            $actividad->NOMBREACTIVIDAD = $request->NOMBREMEDALLA . ' Actividad';
+            $actividad->save();
+        }
+
         /*
         * FIN MEDALLA ASISTENCIA
         */
@@ -169,8 +199,34 @@ class MedallaController extends Controller
         $detalleMedalla = DetalleMedalla::find($id);
         $detalleMedalla->medalla = Medalla::find($detalleMedalla->IDMEDALLA);
 
+        $idDocente = Auth::user()->IDDOCENTE;
+
+        //Obtener las Materias impartidas en Este Ciclo
+        $ciclo = Ciclo::where('ESTAACTIVOCICLO', 1)->first();
+        //Obtener la Materia que Coordina
+        $coordinador = Coordinador::where('IDDOCENTE', $idDocente)->get();
+        $coordinador->each(function($coordinador){
+            $coordinador->materiaImpartida = MateriaImpartida::find($coordinador->IDMATERIAIMPARTIDA);
+        });
+        $Existe = false;
+        foreach($coordinador as $coor){
+            if($coor->materiaImpartida->IDCICLO == $ciclo->id){
+                //Materia que Coordina el Docente
+                $matImp = MateriaImpartida::find($coor->materiaImpartida->id);
+            }
+        }
+        ////
+        $detallePunto = DetallePunto::where('IDMATERIAIMPARTIDA', $matImp->id)->first();
+        $nombreAct = $detalleMedalla->medalla->NOMBREMEDALLA . ' Actividad';
+
+        $actividad = Actividad::where('NOMBREACTIVIDAD', $nombreAct)->first();
+        $detallePunto = DetallePunto::where('id',$actividad->IDDETALLEPUNTO)->where('IDMATERIAIMPARTIDA', $matImp->id)->first();                
+                
+        
         return view('medalla.editar')->with([
-            'detalleMedalla' => $detalleMedalla
+            'detalleMedalla' => $detalleMedalla,
+            'matImp' => $matImp,
+            'detallePunto' => $detallePunto
         ]);
     }
 
@@ -195,7 +251,19 @@ class MedallaController extends Controller
             $detalleMedalla->save();
             //Llenar Medalla editado
             $medalla->fill($request->all());
-        
+
+            $nombreAct = $medalla->NOMBREMEDALLA . ' Actividad';
+            $actividad = Actividad::where('NOMBREACTIVIDAD', $nombreAct)->get();
+            
+            foreach($actividad as $act){
+                $detallePunto = DetallePunto::where('id',$act->IDDETALLEPUNTO)->where('IDMATERIAIMPARTIDA', $request->IDMATERIAIMPARTIDA)->first();                
+                //dd($detallePunto);
+                if($detallePunto){
+                    $detallePunto->PUNTOSACTIVIDAD = $request->PUNTOSACTIVIDAD;
+                    $detallePunto->save();
+                }
+            }
+            
             if($request->file('imgMedalla') != null)
             {
                 $Foto = $request->file('imgMedalla');
@@ -233,16 +301,63 @@ class MedallaController extends Controller
     {
         $detalleMedalla = DetalleMedalla::find($id);
         $medalla = Medalla::find($detalleMedalla->IDMEDALLA);
+        $nombreAct = $medalla->NOMBREMEDALLA . ' Actividad';
 
-        $medallaGanada = MedallaGanada::where('IDDETALLEMEDALLA',$detalleMedalla->id)->first();
-
-        if($medallaGanada != null){
-            Flash::danger("No puede eliminar la Medalla: " . $medalla->NOMBREMEDALLA . " pues esta en uso");
-        }else{
-            $detalleMedalla->delete();
-            $medalla->delete();
-            Flash::danger("Medalla :".$medalla->NOMBREMEDALLA . ' Eliminada con exito');
+        $actividad = Actividad::where('NOMBREACTIVIDAD', $nombreAct)->get();
+        $Estado = 0;
+        foreach($actividad as $act){
+            $detallepuntoactividad = DetallePuntoActividad::where('IDACTIVIDAD', $act->id)->first();
+            if(count($detallepuntoactividad) > 0){
+                $Estado = 1;
+            }
         }
+        //dd($Estado);
+        if($Estado == 1){
+            Flash::danger("Medalla: ".$medalla->NOMBREMEDALLA . ' Ya tiene puntos asignados. No se puede Eliminar');
+        }else{
+            $idDocente = Auth::user()->IDDOCENTE;
+
+            //Obtener las Materias impartidas en Este Ciclo
+            $ciclo = Ciclo::where('ESTAACTIVOCICLO', 1)->first();
+            //Obtener la Materia que Coordina
+            $coordinador = Coordinador::where('IDDOCENTE', $idDocente)->get();
+            $coordinador->each(function($coordinador){
+                $coordinador->materiaImpartida = MateriaImpartida::find($coordinador->IDMATERIAIMPARTIDA);
+            });
+            $Existe = false;
+            foreach($coordinador as $coor){
+                if($coor->materiaImpartida->IDCICLO == $ciclo->id){
+                    //Materia que Coordina el Docente
+                    $matImp = MateriaImpartida::find($coor->materiaImpartida->id);
+                }
+            }
+            $grupo = Grupo::where('IDMATERIAIMPARTIDA',$matImp->id)->get();
+
+            $medallaGanada = MedallaGanada::where('IDDETALLEMEDALLA',$detalleMedalla->id)->first();
+
+            if($medallaGanada != null){
+                Flash::danger("No puede eliminar la Medalla: " . $medalla->NOMBREMEDALLA . " pues esta en uso");
+            }else{
+                $detalleMedalla->delete();
+                $nombreAct = $medalla->NOMBREMEDALLA . ' Actividad';
+                $medalla->delete();
+                
+                $actividad = Actividad::where('NOMBREACTIVIDAD', $nombreAct)->get();
+                
+                foreach($actividad as $act){
+                    $detallePunto = DetallePunto::where('id',$act->IDDETALLEPUNTO)->where('IDMATERIAIMPARTIDA', $matImp->id)->first();                
+                    $tipoActividad = TipoActividad::where('id',$act->IDTIPOACTIVIDAD)->first();
+
+                    $act->delete();
+                    $detallePunto->delete();
+                    $tipoActividad->delete();
+                }
+                Flash::danger("Medalla :".$medalla->NOMBREMEDALLA . ' Eliminada con exito');
+            }                   
+        }
+        
+
+        
        return redirect()->route('medallas.index');
     }
 
@@ -251,6 +366,7 @@ class MedallaController extends Controller
         //dd($request->all());
         $materiaImpartida = MateriaImpartida::find($request->IDMATERIAIMPARTIDA);
         $materia = Materia::find($materiaImpartida->IDMATERIA);
+        $grupo = Grupo::where('IDMATERIAIMPARTIDA',$materiaImpartida->id)->get();
 
         //Creacion de Medallas de Participacion y Asistencia
         /*
@@ -264,7 +380,7 @@ class MedallaController extends Controller
         //Guardar Medalla 
         $medalla->save();
         /*
-        //Detalle de Participacion
+        //Detalle de Asistencia
         */
         $detalleMedalla = new DetalleMedalla;
         $detalleMedalla->IDMATERIAIMPARTIDA = $materiaImpartida->id;
@@ -272,6 +388,26 @@ class MedallaController extends Controller
         $detalleMedalla->CANTIDADMINIMAPUNTOS = 50;
         //Guardar Detalle Medalla
         $detalleMedalla->save();
+
+        foreach($grupo as $g){
+            //DETALLE ACTIVIDAD
+            $detallePunto = new DetallePunto();
+            $detallePunto->IDTIPOPUNTO = 2;
+            $detallePunto->IDMATERIAIMPARTIDA = $materiaImpartida->id;
+            $detallePunto->ESTAACTIVOPUNTO = 1;
+            $detallePunto->PUNTOSACTIVIDAD = 5;
+            $detallePunto->save();
+
+            //ACTIVIDAD DE ASISTENCIA
+            $actividad = new Actividad();
+            $actividad->IDTIPOACTIVIDAD = 1;
+            $actividad->IDDETALLEPUNTO = $detallePunto->id;
+            $actividad->IDGRUPO = $g->id;
+            $actividad->NOMBREACTIVIDAD = 'Asistencia';
+            $actividad->save();
+        }
+        
+
         /*
         * FIN MEDALLA ASISTENCIA
         */
@@ -292,9 +428,27 @@ class MedallaController extends Controller
         $detalleMedalla = new DetalleMedalla;
         $detalleMedalla->IDMATERIAIMPARTIDA = $materiaImpartida->id;
         $detalleMedalla->IDMEDALLA = $medalla->id;
-        $detalleMedalla->CANTIDADMINIMAPUNTOS = 2;
+        $detalleMedalla->CANTIDADMINIMAPUNTOS = 5;
         //Guardar Detalle Medalla
         $detalleMedalla->save();
+
+        foreach($grupo as $g){
+            //DETALLE ACTIVIDAD
+            $detallePunto = new DetallePunto();
+            $detallePunto->IDTIPOPUNTO = 2;
+            $detallePunto->IDMATERIAIMPARTIDA = $materiaImpartida->id;
+            $detallePunto->ESTAACTIVOPUNTO = 1;
+            $detallePunto->PUNTOSACTIVIDAD = 1;
+            $detallePunto->save();
+
+            //ACTIVIDAD DE PARTICIPACION
+            $actividad = new Actividad();
+            $actividad->IDTIPOACTIVIDAD = 2;
+            $actividad->IDDETALLEPUNTO = $detallePunto->id;
+            $actividad->IDGRUPO = $g->id;
+            $actividad->NOMBREACTIVIDAD = 'Participacion';
+            $actividad->save();
+        }
         /*
         * FIN MEDALLA PARTICIPACION
         */
